Do not mask parse erros with default values
This fixes a longstanding problem of ignoring parse errors
when the variable the user passed a malformed value for happens
to have a default value.

There are 2 obvious ways to fix that: either localize the improved
parsing logic to the `VarF` constructor or move variable lookup
to the top level (where default values are dealt with). The second
method has been chosen since it seems close to what the user's
assumptions would be and the code appears to be clearer. An unfortunate
consequence is that the `AsUnset` constraint is propagated upwards
to the `parse*` functions, although it shouldn't matter in practice
as every reasonable error type must deal with unset variables somehow
anyway.

Fixes #8